conversation opt

+++ internal/repository/base.go
@@ -4,7 +4,7 @@ import "gorm.io/gorm"
 
 type Generic[T any] interface {
 	Create(db *gorm.DB, entity *T) error
-	GetById(db *gorm.DB, id int64) (*T, error)
+	GetById(db *gorm.DB, id int) (*T, error)
 	Update(db *gorm.DB, entity *T) error
 	List(db *gorm.DB, page, pageSize int) ([]T, error)
 }
@@ -15,7 +15,7 @@ func (g GenericImpl[T]) Create(db *gorm.DB, entity *T) error {
 	return db.Create(entity).Error
 }
 
-func (g GenericImpl[T]) GetById(db *gorm.DB, id int64) (*T, error) {
+func (g GenericImpl[T]) GetById(db *gorm.DB, id int) (*T, error) {
 	var entity T
 	err := db.First(&entity, id).Error
 	return &entity, err
